pyproject.toml
[project]
name = "chefmatch"
version = "0.1.0"
description = ""
authors = [
    {name = "Camila",email = "milarodcar@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "djangorestframework (>=3.15.2,<4.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "streamlit (>=1.42.2,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
